+++ parallel_prog/lib/math.cc
@@ -2,8 +2,10 @@
 
 #include <cmath>
 
-double L1Norm(const std::vector<double> a, const std::vector<double> b) {
-	double sum = 0;
+long double L1Norm(
+		const std::vector<long double> a,
+		const std::vector<long double> b) {
+	long double sum = 0;
 	for (int i = 0; i < a.size(); i++) {
 		sum += std::abs(a[i] - b[i]);
 	}
