+++ parallel_prog/src/omp.cc
@@ -25,12 +25,12 @@ DEFINE_double(eps, 1e-7, "Computation precision");
 void ParallelAddPagesPr(
 		const std::vector<pr::Page> &pages,
 		const std::vector<int> &out_link_cnts,
-		const std::vector<double> &old_pr,
-		std::vector<double> &new_pr) {
+		const std::vector<long double> &old_pr,
+		std::vector<long double> &new_pr) {
 	int i;
 #pragma omp parallel for
 	for (i = 0; i < pages.size(); i++) {
-		double sum = 0;
+		long double sum = 0;
 		for (int from_page : pages[i].in_links()) {
 			sum += old_pr[from_page] / out_link_cnts[from_page];
 		}
@@ -55,8 +55,8 @@ int main(int argc, char *argv[]) {
 	timer.StopAndReport("Reading pages");
 
 	std::vector<int> dangling_pages = ExploreDanglingPages(out_link_cnts);
-	std::vector<double> pr = InitPr(page_cnt);
-	std::vector<double> old_pr(page_cnt);
+	std::vector<long double> pr = InitPr(page_cnt);
+	std::vector<long double> old_pr(page_cnt);
 
 	timer.Start();
 	bool go_on = true;
@@ -68,7 +68,7 @@ int main(int argc, char *argv[]) {
 		AddDanglingPagesPr(FLAGS_damping_factor, dangling_pages, old_pr, pr);
 		AddRandomJumpsPr(FLAGS_damping_factor, pr);
 
-		double err = L1Norm(pr, old_pr);
+		long double err = L1Norm(pr, old_pr);
 		LOG(INFO) << "Error " << std::setprecision(10) << std::fixed << err
 			<< " at step " << step + 1;
 		go_on = err > FLAGS_eps;
